+++ src/__tests__/BaseLayout.test.tsx
@@ -2,19 +2,18 @@ import { describe, expect, it } from 'vitest';
 import { render, screen } from '@testing-library/react';
 import BaseLayout from '../layouts/Base';
 import { BrowserRouter } from 'react-router';
-import ThemeProvider from '../components/theme/ThemeProvider';
-import { Provider } from 'react-redux';
-import store from '../store';
+import ThemeProvider from '../providers/theme';
+import ReduxProvider from '../providers/redux';
 
 describe('BaseLayout', () => {
   it('renders base layout with header', () => {
     render(
       <BrowserRouter>
-        <Provider store={store}>
+        <ReduxProvider>
           <ThemeProvider>
             <BaseLayout />
           </ThemeProvider>
-        </Provider>
+        </ReduxProvider>
       </BrowserRouter>
     );
