+++ src/__tests__/HomePage.test.tsx
@@ -2,16 +2,15 @@ import { describe, expect, it } from 'vitest';
 import { render, screen } from '@testing-library/react';
 import HomePage from '../pages/HomePage';
 import { BrowserRouter } from 'react-router';
-import { Provider } from 'react-redux';
-import store from '../store';
+import ReduxProvider from '../providers/redux';
 
 describe('HomePage', () => {
   it('renders home page', () => {
     render(
       <BrowserRouter>
-        <Provider store={store}>
+        <ReduxProvider>
           <HomePage />
-        </Provider>
+        </ReduxProvider>
       </BrowserRouter>
     );
