--- src/components/theme/types.ts
@@ -1,10 +1,0 @@
-export const themeList = {
-  light: 'light',
-  dark: 'dark',
-} as const;
-
-export type Theme = keyof typeof themeList;
-export type ComplexTheme = {
-  theme: Theme | null;
-  setTheme: React.Dispatch<React.SetStateAction<Theme | null>>;
-};
