+++ src/constants/theme.ts
@@ -1,0 +1,4 @@
+export const themeList = {
+  light: 'light',
+  dark: 'dark',
+} as const;
