+++ src/contexts/theme.ts
@@ -1,4 +1,4 @@
 import { createContext } from 'react';
-import type { ComplexTheme } from './types';
+import type { ComplexTheme } from '../types';
 
 export const ThemeContext = createContext<ComplexTheme | null>(null);
