+++ src/providers/redux.tsx
@@ -1,0 +1,12 @@
+import { Provider } from 'react-redux';
+import store from '../store';
+
+type Props = {
+  children: React.ReactNode;
+};
+
+const ReduxProvider = ({ children }: Props) => {
+  return <Provider store={store}>{children}</Provider>;
+};
+
+export default ReduxProvider;
