+++ src/types/index.ts
@@ -1,3 +1,5 @@
+import { themeList } from '../constants/theme';
+
 export type Character = {
   id: number;
   name: string;
@@ -7,11 +9,9 @@ export type Character = {
   image: string;
   url: string;
 };
-
 export type CharacterSummary = Pick<Character, 'id' | 'name' | 'url'> & {
   description: string;
 };
-
 export type CharacterSummaryWithChecked = CharacterSummary & {
   checked: boolean;
 };
@@ -21,4 +21,10 @@ export type Info = {
   pages: number;
   next: string | null;
   prev: string | null;
+};
+
+export type Theme = keyof typeof themeList;
+export type ComplexTheme = {
+  theme: Theme | null;
+  setTheme: React.Dispatch<React.SetStateAction<Theme | null>>;
 };
